+++ src/cmd_rgr/mod.rs
@@ -10,4 +10,5 @@ pub mod pl_2rmp;
 pub mod prop;
 pub mod replace;
 pub mod runlist;
+pub mod select;
 pub mod sort;