+++ tests/cli_rgr_tsv.rs
@@ -145,3 +145,48 @@ fn command_filter() -> anyhow::Result<()> {
 
     Ok(())
 }
+
+#[test]
+fn command_select() -> anyhow::Result<()> {
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("select")
+        .arg("tests/rgr/ctg.tsv")
+        .arg("-H")
+        .arg("-f")
+        .arg("6,1")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 4);
+    assert_eq!(
+        stdout.lines().next().unwrap().split('\t').count(),
+        2,
+        "field count"
+    );
+    assert!(stdout.contains("length\tID"));
+    assert!(stdout.contains("130218\tctg:I:2"));
+
+    let mut cmd = Command::cargo_bin("rgr")?;
+    let output = cmd
+        .arg("select")
+        .arg("tests/rgr/ctg.tsv")
+        .arg("-H")
+        .arg("-f")
+        .arg("ID,1")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 4);
+    assert_eq!(
+        stdout.lines().next().unwrap().split('\t').count(),
+        2,
+        "field count"
+    );
+    assert!(stdout.contains("ID\tID"));
+    assert!(stdout.contains("ctg:I:2\tctg:I:2"));
+
+    Ok(())
+}